clean up

+++ day2/part1/part1.rs
@@ -3,10 +3,10 @@ fn main() {
     include_str!("../input")
     .split("\r\n")
     .for_each(|x| {
-        let opponent =  x.chars().nth(0).unwrap() ;
-        let player =  x.chars().nth(2).unwrap();
-        let residue = ((player as i64 - opponent as i64) % 3) as f64;
-        output+=(4.5* residue * residue - 10.5 * residue) as i64 + 6 + (player as i64 - 87);   
+        let opponent =  x.chars().nth(0).unwrap() as i64;
+        let player =  x.chars().nth(2).unwrap() as i64;
+        let residue = ((player - opponent) % 3) as f64;
+        output+=(4.5* residue * residue - 10.5 * residue) as i64 + 6 + (player - 87);   
     });
     println!("{}",output);
 }
